test(core): split skill lifecycle test responsibilities
